102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,18 +5,18 @@
 */
 void print_fibonacci(void)
 {
-int term1 = 1;
-int term2 = 2;
-int fibn = 0;
+long long int term1 = 1;
+long long int term2 = 2;
+long long int fibn = 0;
 int i;
 int ln = 50;
 
-printf("%d, %d", term1, term2);
+printf("%lld, %lld", term1, term2);
 
 for (i = 3; i <= ln; i++)
 {
 fibn = term1 + term2;
-printf(", %d", fibn);
+printf(", %lld", fibn);
 term1 = term2;
 term2 = fibn;
 }
